Phase 2a: Extract build subcommand into commands/build.py
Moves run_build_command and its imports out of commands/__init__.py
into a focused module. The package re-exports the name for backwards
compatibility ('from rushti.commands import run_build_command' continues
to work).

This is the first of several per-subcommand extractions. Pattern:
1. New module gets only the imports it actually needs
2. Lazy import of CONFIG from cli.py preserved (cycle break)
3. __init__.py drops the function body and re-exports from the new module
4. Now-unused imports in __init__.py are cleaned up

commands/__init__.py shrinks from 2108 to ~2000 lines. Full unit suite
still passes (637 tests).
